Create Maximum consecutive one’s in a binary array
Given an array of only 0’s and 1’s. You have to find out the maximum number of consecutive ones in the array.

Input Format
The first line takes a number N representing the size of the array. The second line has N integers representing the elements of an array.

Constraints
The length of the array must always be less than 10,000.
The array only contains 0 and 1.

Output Format
The maximum length of consecutive ones.

Sample Input
10
0 0 1 0 1 1 0 0 1 0
Sample Output
2
Explanation
Input: { 0, 0, 1, 0, 1, 1, 0, 0, 1, 0}
Output: 2
The maximum length consecutive zeroes is 2. Hence this is the ouput.