Broken Keyboard
You are a noob gamer. You often lose while playing online games and end up punching your keyboard a lot.

Now you want to create a new account. But you notice that most of the keys don't work properly. So you have to press them for a long time. Due to this problem, a character will be typed 1 or more times.

You examine the typed characters of the keyboard. Output True if it is possible that it was your name, with some characters (possibly none) being long pressed, otherwise output false.

Input Format
2 strings, one with the correct name followed by another with long pressed name.

Constraints
1 <= name.length(), typed.length() <= 1000
Both strings consist of only lowercase English letters.

Output Format
true or false, depending on whether the second string is the long pressed version of the correct name

Sample Input
tushar
tuuushaarr
Sample Output
true
Explanation
In the second string, 'u', 'a' and 'r' are pressed multiple times and are in the correct sequence.