Create Monotonous array.java
Ram loves monotonic array. He wants to find out whether the given array is monotonic or not.

An array is monotonic if it is either monotone increasing or monotone decreasing.

An array nums is monotone increasing if for all i <= j, nums[i] <= nums[j]. An array nums is monotone decreasing if for all i <= j, nums[i] >= nums[j].

You are given an integer array nums, Print True if the given array is monotonic, or False otherwise.

Input Format
First Line contains an integer N (size of the array) .
Second Line contains an array.

Constraints
1 <= N <= 10000
-1000 <= nums[i] <= 1000

Output Format
Print "True" or "False".

Sample Input
5 
1 4 2 5 3
Sample Output
False
Explanation
First Array is increasing then it is decreasing .
So the given array is not monotonous.