Create  diameter of a binary tree.cpp
The diameter of a tree (sometimes called the width) is the number of nodes on the longest path between two end nodes. The diagram below shows two trees each with diameter nine, the leaves that form the ends of the longest path are shaded (note that there is more than one path in each tree of length nine, but no path longer than nine nodes).

Input Format
Input tree is given in the form of level order traversal, -1 means child is NULL.

Constraints
1<=nodes.data<=10^5

Output Format
An integer representing the diameter of the tree.

Sample Input
1 2 3 4 5 -1 -1 -1 -1 -1 -1 
Sample Output
4
Explanation
None.