Create Reverse the linked list in between..cpp
There is a family on the trip of N members and they have already decided that they will be maitaining their order in the manner just like the linked list and at certain point due to thunderstorm they got panic and also their order got disturbed . they know that their order is so disturbed that if the whole members list from left point to the right point is reversed then their order will be maintianed again.

Input Format
First line contains the number of member in the family--N Second line contains the N member of the family. Third line contains the value of the left pointer and the right pointer from where we have to reverse the list to get the same order back.

Constraints
1 <= n <= 500
-500 <= Node.val <= 500
1 <= left <= right <= n

Output Format
You have to print the order after reversing the linked list from the given points.

Sample Input
5
1 2 3 4 5
2 4
Sample Output
1 4 3 2 5 
Explanation
left pointer points to the node '4' and the right pointer points to the node '2'. After reversing this much part of the list the order will be maintained.