Create Design Linked List.cpp
Complete the below functions as decribed:
int get(int index)Get the value of the indexth node in the linked list. If the index is invalid, return -1.
void addAtHead(int val) Add a node of value val before the first element of the linked list. After the insertion, the new node will be the first node of the linked list.
void addAtTail(int val) Append a node of value val as the last element of the linked list.
void addAtIndex(int index, int val) Add a node of value val before the indexth node in the linked list. If index equals the length of the linked list, the node will be appended to the end of the linked list. If index is greater than the length, the node will not be inserted.
void deleteAtIndex(int index) Delete the indexth node in the linked list, if the index is valid.

Input Format
First line contains how many functions you have to perform and according to that you have to take values.You should make the function calls for their working as the same as mentioned in the above problem.

Constraints
0 <= index, val <= 1000

Output Format
If the output is needed then you will print in space separates manner.

Sample Input
6
addAtHead
1
addAtTail
3
addAtIndex
1 2
get
1
deleteAtIndex
1
get
1
Sample Output
2 3
Explanation
addAtHead(1);
addAtTail(3);
addAtIndex(1, 2); // linked list becomes 1->2->3
get(1); // return 2
deleteAtIndex(1); // now the linked list is 1->3
get(1);