Update Count Smaller.java
You are given an integer array nums and you have to return a new counts array. The counts array has the property where counts[i] is the number of smaller elements to the right of nums[i].

Input Format
The first line of input will contain integer N, denoting the size of the array. Second line will contain N space-separated integers, denoting elements of the array.

Constraints
1 <= N <= 100000

Output Format
Print the output array in space-separated manner.

Sample Input
4
5 2 6 1
Sample Output
2 1 1 0 
Explanation
To the right of 5 there are 2 smaller elements (2 and 1).
To the right of 2 there is only 1 smaller element (1).
To the right of 6 there is 1 smaller element (1).
To the right of 1 there is 0 smaller element.