Create Repeating Substring.java
Konark is a very hardworking boy. He is the topper of his class. Today his strict teacher Dr. Pawan gave him a homework to find the length of longest substring in which the characters must not repeat. Help Konark in his homework.

Input Format
Input the string

Constraints
0 <= s.length <= 5 * 104
String consists of English letters, digits, symbols and spaces.

Output Format
Return the length of longest substring

Sample Input
abcabcbb
Sample Output
3
Explanation
The answer is "abc", with the length of 3.